requirements.txt
Django==5.0.0
pymongo==4.6.1
python-dotenv==1.0.0
django-apscheduler==0.6.2
selenium==4.18.1
beautifulsoup4==4.12.3
